refactor common code used by both CI Visibility command
