+++ tracer/src/Datadog.Trace.Tools.Runner/RunHelper.cs
@@ -10,19 +10,18 @@ namespace Datadog.Trace.Tools.Runner
 {
     internal static class RunHelper
     {
-        public static bool TryGetEnvironmentVariables(ApplicationContext applicationContext, InvocationContext invocationContext, RunSettings settings, out Dictionary<string, string> profilerEnvironmentVariables)
+        public static bool TryGetEnvironmentVariables(ApplicationContext applicationContext, InvocationContext invocationContext, CommonTracerSettings settings, out Dictionary<string, string> profilerEnvironmentVariables)
         {
             return TryGetEnvironmentVariables(applicationContext, invocationContext, settings, Utils.CIVisibilityOptions.None, out profilerEnvironmentVariables);
         }
 
-        public static bool TryGetEnvironmentVariables(ApplicationContext applicationContext, InvocationContext invocationContext, RunSettings settings, Utils.CIVisibilityOptions ciVisibilityOptions, out Dictionary<string, string> profilerEnvironmentVariables)
+        public static bool TryGetEnvironmentVariables(ApplicationContext applicationContext, InvocationContext invocationContext, CommonTracerSettings settings, Utils.CIVisibilityOptions ciVisibilityOptions, out Dictionary<string, string> profilerEnvironmentVariables)
         {
             profilerEnvironmentVariables = Utils.GetProfilerEnvironmentVariables(
                 invocationContext,
                 applicationContext.RunnerFolder,
                 applicationContext.Platform,
                 settings,
-                reducePathLength: false,
                 ciVisibilityOptions: ciVisibilityOptions);
 
             if (profilerEnvironmentVariables is null)
@@ -30,15 +29,18 @@ namespace Datadog.Trace.Tools.Runner
                 return false;
             }
 
-            var additionalEnvironmentVariables = invocationContext.ParseResult.GetValueForOption(settings.AdditionalEnvironmentVariables);
-
-            if (additionalEnvironmentVariables != null)
+            if (settings is RunSettings runSettings)
             {
-                foreach (var env in additionalEnvironmentVariables)
-                {
-                    var (key, value) = ParseEnvironmentVariable(env);
+                var additionalEnvironmentVariables = invocationContext.ParseResult.GetValueForOption(runSettings.AdditionalEnvironmentVariables);
 
-                    profilerEnvironmentVariables[key] = value;
+                if (additionalEnvironmentVariables != null)
+                {
+                    foreach (var env in additionalEnvironmentVariables)
+                    {
+                        var (key, value) = ParseEnvironmentVariable(env);
+
+                        profilerEnvironmentVariables[key] = value;
+                    }
                 }
             }
